Genestealer Cult, Showing Weapons Options
Genestealer Cult, Showing Weapons Options

+++ ArmyStarter.Blazor/Provider/AttackingModel.cs
@@ -8,7 +8,7 @@
 
         public int Strength { get; set; }
 
-        public int Attacks { get; set; }
+        public decimal Attacks { get; set; }
 
         public int AP { get; set; }
     }
